Updated variable type for addPost to int. Able to delete posts.

+++ src/main/java/com/example/springbootspillit/repository/PostRepository.java
@@ -3,6 +3,6 @@ package com.example.springbootspillit.repository;
 import com.example.springbootspillit.model.Posts;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Posts, Long > {
+public interface PostRepository extends CrudRepository<Posts, Integer > {
 
 }
